fix(ibc): move escrow accounts migration to v11 upgradeHandler
