--- app/upgrades/v10_1/constants.go
@@ -1,12 +1,0 @@
-package v101
-
-const (
-	// UpgradeName is the shared upgrade plan name for mainnet
-	UpgradeName = "v10.1.0"
-	// UpgradeInfo defines the binaries that will be used for the upgrade
-	UpgradeInfo = `'{"binaries":{"darwin/arm64":"https://github.com/evmos/evmos/releases/download/v10.1.0/evmos_10.1.0_Darwin_arm64.tar.gz","darwin/amd64":"https://github.com/evmos/evmos/releases/download/v10.1.0/evmos_10.1.0_Darwin_amd64.tar.gz","linux/arm64":"https://github.com/evmos/evmos/releases/download/v10.1.0/evmos_10.1.0_Linux_arm64.tar.gz","linux/amd64":"https://github.com/evmos/evmos/releases/download/v10.1.0/evmos_10.1.0_Linux_amd64.tar.gz","windows/x86_64":"https://github.com/evmos/evmos/releases/download/v10.1.0/evmos_10.1.0_Windows_x86_64.zip"}}'`
-
-	// at the time of this migration, on mainnet, channels 0 to 36 were open
-	// so this migration covers those channels only
-	openChannels = 36
-)
