+++ app/upgrades/v11/upgrades_test.go
@@ -1,4 +1,4 @@
-package v101_test
+package v11_test
 
 import (
 	"fmt"
@@ -11,7 +11,7 @@ import (
 	"github.com/evmos/ethermint/crypto/ethsecp256k1"
 	feemarkettypes "github.com/evmos/ethermint/x/feemarket/types"
 	"github.com/evmos/evmos/v10/app"
-	v10 "github.com/evmos/evmos/v10/app/upgrades/v10_1"
+	v11 "github.com/evmos/evmos/v10/app/upgrades/v11"
 	"github.com/stretchr/testify/require"
 	"github.com/tendermint/tendermint/crypto/tmhash"
 	tmproto "github.com/tendermint/tendermint/proto/tendermint/types"
@@ -73,7 +73,7 @@ func TestMigrateEscrowAcc(t *testing.T) {
 	setupEscrowAccounts(app, ctx, existingAccounts)
 
 	// Run migrations
-	v10.MigrateEscrowAccounts(ctx, app.AccountKeeper)
+	v11.MigrateEscrowAccounts(ctx, app.AccountKeeper)
 
 	// check account types for channels 0 to 36
 	for i := 0; i <= 36; i++ {
